+++ lib/puppet-check/tasks.rb
@@ -21,7 +21,7 @@ class PuppetCheck::Tasks < ::Rake::TaskLib
 
       desc 'Execute RSpec and RSpec-Puppet tests'
       RSpec::Core::RakeTask.new(:spec) do |task|
-        rspec_puppet_setup
+        self.class.rspec_puppet_setup
         # generate tasks for all recognized directories
         task.pattern = '**/{classes, defines, facter, functions, hosts, puppet, unit, types}/**/*_spec.rb'
       end
@@ -34,20 +34,39 @@ class PuppetCheck::Tasks < ::Rake::TaskLib
   end
 
   # prepare the directories for rspec-puppet testing
-  def rspec_puppet_setup
-    # leave method immediately if there is no rspec-puppet installed
-    begin
-      require 'rspec-puppet/setup'
-    rescue LoadError
-      return
-    end
-
-    # executes rspec::puppet::setup in every module directory to ensure module spec directories are configured correctly
+  def self.rspec_puppet_setup
     # ensure this method does not do anything inside module dependencies
-    specdirs = Dir.glob('**/spec').reject! { |dir| dir =~ /fixtures/ }
+    specdirs = Dir.glob('**/spec').reject { |dir| dir =~ /fixtures/ }
+    return if specdirs.class.to_s == 'NilClass'
+
+    # setup fixtures for rspec-puppet testing
     specdirs.each do |specdir|
       Dir.chdir(specdir + '/..')
-      RSpec::Puppet::Setup.run
+
+      # skip to next specdir if it does not seem like a puppet module
+      next unless File.directory?('manifests')
+
+      # grab the module name from the directory name of the module
+      module_name = File.basename(Dir.pwd)
+
+      # create all the necessary fixture dirs that are missing
+      ['spec/fixtures', 'spec/fixtures/manifests', 'spec/fixtures/modules', "spec/fixtures/modules/#{module_name}"].each do |dir|
+        FileUtils.mkdir(dir) unless File.directory?(dir)
+      end
+
+      # create empty site.pp if missing
+      FileUtils.touch('spec/fixtures/manifests/site.pp') unless File.file?('spec/fixtures/manifests/site.pp')
+
+      # symlink over everything the module needs for compilation
+      %w(hiera.yaml data hieradata functions manifests lib files templates).each do |file|
+        FileUtils.ln_s("../../../../#{file}", "spec/fixtures/modules/#{module_name}/#{file}") if File.exist?(file)
+      end
+
+      # create spec_helper if missing
+      next if File.file?('spec/spec_helper.rb')
+      File.open('spec/spec_helper.rb', 'w') do |file|
+        file.puts "require 'rspec-puppet/spec_helper'\n"
+      end
     end
   end
 end